- Shafe - In Arsenal and LMS if you happen to blow yourself up while making the winning kill you will still win assuming you died last.  The game will not get stuck looking for a winner.
- Shafe - First Strike Code Now Works.. This was important in arsenal and lms.

- Shafe - in Arsenal & LMS you are spectator when you connect, thus not having a round go 
on for 15 minutes, down to 2, then someone connects now it's unfairly 3.. the new people 
will need to wait to next round to play.

- Shafe - Did a bit of work to the connect screen.  Progress bar is still a bit screwy.

- Shafe - Added Callvote and Choose Class to the ingame menu but they dont do anything 
yet.

- Shafe - Add Bot, Remove Bot, and Restart Map never appear in the ingame menu when 
connected to a server. 

- Shafe - Added in last man standing game mode (which right now is pretty bland and basic) 
the code was mostly there, being used for Arsenal.  Last Man Standing is g_GameMode 2.  We 
should come up with a better name.  LMS in Trep is more similar to Gladiator in EF than 
Elimination or other Last Man Standing Game Types.