- Shafe - Started adding GameMode 3.. No details yet.. Shhhhhh.. it's a secret
- Shafe - Started Incorporating HUD.. Not done yet though

- Shafe - Added in breakable Objects functions.. This is from a tutorial on code3arena, so I dont know if it works.. but here is the notes from the tutorial as well as some of my own.
Mappers need to first, make it a func_door(anything will do) then press "n" and change the func_door to func_breakable, or optionally edit the "entities.def" file and add in func_breakable, so you can just select it from the normal drop down list in Q3Radiant. They also need to set the life of the glass by pressing "n" while selecting the entity and in the key box type in "health" and in the value box type in whatever NUMBER of life it should have. If you dont it defaults to 5 life. 
Your mapper needs to add the ifx.shader into their base/scripts directory 
You will want he/she to edit the shaderlist.txt in their base/scripts directory and add at the VERY bottom ifx
Your mapper needs to have done that so upon compile it does the lightmap and everything correctly.

So this is all from a code3arena tutorial... I imagine the shader only has to do with rendering the glass model correctly.  If you dont assign a glass model (or whatever object it is that is breakable).  the following will be used:
models/breakables/glass01.md3
models/breakables/glass02.md3
models/breakables/glass03.md3